backend/pyproject.toml
[project]
name = "discboard"
version = "0.1.0"
description = "Mission Control for Discord"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlmodel>=0.0.16",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "sse-starlette>=1.8.0",
]

[project.optional-dependencies]
discord = [
    "discord.py>=2.4.0",
    "python-dotenv>=1.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]